Move DTO to inside the class
